Introduce a `Reduction` base; fix `reduce_sum` dtype promotion.
`ReduceMax`/`Min`/`Sum`/`Prod`/`All`/`LogSumExp` and `Argmax` all
`extends ReduceMean` (or a leaf of it) to borrow its axis-collapse
shape machinery — code-reuse-not-is-a (wala/ML#514). Extract that
machinery into an abstract `Reduction` base whose default dtype
behavior *preserves* the input dtype (as most reductions do); the
`Reduce*` ops and `Argmax` extend it directly, and `Argmin` keeps
`extends Argmax` (an index reduction genuinely like argmax).

This also fixes a latent bug the audit surfaced: `ReduceMean` promotes
integer inputs to `float32` (correct — the mean of ints is a float),
but `ReduceSum` inherited that promotion via `extends ReduceMean`, so
`tf.reduce_sum` of an `int32` was mistyped `float32`. With `Reduction`
preserving the input dtype and `ReduceMean` overriding to promote,
`reduce_sum` now preserves `int32`. `ReduceMax`/`Prod`/`LogSumExp`,
whose explicit preserve-overrides are now the default, are slimmed to
thin subclasses. Added `testReduceSumInt` to guard the corrected
behavior. Full suite green (970 tests).
