+++ com.ibm.wala.cast.python.ml/source/com/ibm/wala/cast/python/ml/client/ReduceMax.java
@@ -1,31 +1,18 @@
 package com.ibm.wala.cast.python.ml.client;
 
-import com.ibm.wala.cast.python.ml.types.TensorFlowTypes.DType;
 import com.ibm.wala.ipa.callgraph.propagation.PointsToSetVariable;
-import com.ibm.wala.ipa.callgraph.propagation.PropagationCallGraphBuilder;
-import java.util.Set;
 
 /**
- * Generator for {@code tf.math.reduce_max}. Reuses {@link ReduceMean}'s axis-collapse / keepdims
- * shape inference, but overrides {@link #getDTypes(PropagationCallGraphBuilder)} to inherit the
- * input's dtype directly (no {@code int → float32} promotion that {@link ReduceMean} applies for
- * its mean-specific runtime semantics). See wala/ML#449 (Tier 3).
+ * Generator for {@code tf.math.reduce_max}. The axis-collapse / keepdims shape inference and the
+ * dtype-preserving default both come from {@link Reduction}. See wala/ML#449 (Tier 3).
  *
  * @see <a
  *     href="https://www.tensorflow.org/api_docs/python/tf/math/reduce_max">tf.math.reduce_max</a>
  * @author <a href="mailto:khatchad@hunter.cuny.edu">Raffi Khatchadourian</a>
  */
-public class ReduceMax extends ReduceMean {
+public class ReduceMax extends Reduction {
 
   public ReduceMax(PointsToSetVariable source) {
     super(source);
-  }
-
-  @Override
-  protected Set<DType> getDTypes(PropagationCallGraphBuilder builder) {
-    int inputValNum =
-        this.getArgumentValueNumber(
-            builder, Parameters.INPUT_TENSOR.getIndex(), Parameters.INPUT_TENSOR.getName(), false);
-    return this.getDTypes(builder, inputValNum);
   }
 }
