+++ com.ibm.wala.cast.python.test/data/tf2_test_reduce_sum_int.py
@@ -1,0 +1,15 @@
+import tensorflow as tf
+
+# Guards that `tf.reduce_sum` preserves an integer input's dtype (no float32 promotion). Unlike
+# `tf.reduce_mean`, summing integers yields an integer (wala/ML#514).
+
+
+def f(x):
+    pass
+
+
+a = tf.constant([1, 2, 3], dtype=tf.int32)
+r = tf.reduce_sum(a)
+assert r.dtype == tf.int32
+assert r.shape == ()
+f(r)
